feat paginate

+++ tests/models/user.go
@@ -3,7 +3,7 @@ package models
 type User struct {
 	ID        *string `bson:"_id,omitempty"`
 	Name      *string `bson:"name,omitempty"`
-	Age       *uint16 `bson:"age,omitempty"`
-	CreatedAt *uint64 `bson:"created_at,omitempty"`
-	UpdatedAt *uint64 `bson:"updated_at,omitempty"`
+	Age       *int    `bson:"age,omitempty"`
+	CreatedAt *int64  `bson:"created_at,omitempty"`
+	UpdatedAt *int64  `bson:"updated_at,omitempty"`
 }
